refactor: rename playground -> pg
