+++ src/app/(playground)/pg/bottom-sheet/_components/image-radio/index.tsx
@@ -1,0 +1,8 @@
+import { ImageRadio } from "~/app/(playground)/pg/bottom-sheet/_components/image-radio/container";
+import { ImageRadioOption } from "~/app/(playground)/pg/bottom-sheet/_components/image-radio/option";
+
+const ImageRadioComponent = Object.assign(ImageRadio, {
+  Option: ImageRadioOption,
+});
+
+export default ImageRadioComponent;
