+++ src/app/(playground)/pg/map/_components/AddressSearchButton/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useKakaoAddress } from "~/app/(playground)/playground/map/_components/KakaoAddressContext";
+import { useKakaoAddress } from "~/app/(playground)/pg/map/_components/KakaoAddressContext";
 
 export default function AddressSearchButton() {
   const { setCoordinate } = useKakaoAddress();
