+++ src/app/(playground)/pg/test/page.tsx
@@ -4,9 +4,11 @@ import {
   QueryClient,
 } from "@tanstack/react-query";
 import { cache } from "react";
-import { ALink, AMain } from "~/app/(playground)/playground/inner-tools";
-import InvitationResponseList from "~/app/(playground)/playground/invitation-response/invitation-response-list";
-import { getAllInvitationResponses } from "~/lib/db/schema/invitation_response.query";
+import { ALink, AMain } from "~/app/(playground)/pg/inner-tools";
+import TestEditDialog from "~/app/(playground)/pg/test/test-edit-dialog";
+import TestForm from "~/app/(playground)/pg/test/test-form";
+import TestList from "~/app/(playground)/pg/test/test-list";
+import { getTestWithTestJobCnt } from "~/lib/db/schema/test.query";
 
 const getQueryClient = cache(() => new QueryClient());
 
@@ -14,15 +16,17 @@ export default async function Page() {
   const queryClient = getQueryClient();
 
   await queryClient.prefetchQuery({
-    queryKey: ["invitationResponses"],
-    queryFn: getAllInvitationResponses,
+    queryKey: ["tests"],
+    queryFn: getTestWithTestJobCnt,
   });
 
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
       <AMain>
         <ALink href="/playground">playground</ALink>
-        <InvitationResponseList />
+        <TestForm />
+        <TestList />
+        <TestEditDialog />
       </AMain>
     </HydrationBoundary>
   );
