--- src/app/(playground)/playground/bottom-sheet/_components/image-radio/index.tsx
@@ -1,8 +1,0 @@
-import { ImageRadio } from "~/app/(playground)/playground/bottom-sheet/_components/image-radio/container";
-import { ImageRadioOption } from "~/app/(playground)/playground/bottom-sheet/_components/image-radio/option";
-
-const ImageRadioComponent = Object.assign(ImageRadio, {
-  Option: ImageRadioOption,
-});
-
-export default ImageRadioComponent;
